clientesyproducot

+++ ms-fyv/src/main/java/com/example/msfyv/controller/FacturaController.java
@@ -44,12 +44,12 @@ public class FacturaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Factura> listById(@PathVariable(required = true) Double id){
+    public ResponseEntity<Factura> listById(@PathVariable(required = true) Integer id){
         return ResponseEntity.ok().body(facturaService.listarPorId(id).get());
     }
 
     @DeleteMapping("/{id}")
-    public String deleteById(@PathVariable(required = true) Double id){ facturaService.eliminarPorId(id);
+    public String deleteById(@PathVariable(required = true) Integer id){ facturaService.eliminarPorId(id);
         return "Eliminado Correctamente :3";
     }
