+++ ms-fyv/src/main/java/com/example/msfyv/service/FacturaService.java
@@ -13,8 +13,8 @@ public interface FacturaService {
 
     public Factura actualizar(Factura factura);
 
-    public Optional<Factura> listarPorId(Double id);
+    public Optional<Factura> listarPorId(Integer id);
 
-    public void eliminarPorId(Double id);
+    public void eliminarPorId(Integer id);
 
 }
